Fix Gfycat strategy
