+++ .circleci/config.yml
@@ -5,7 +5,7 @@ jobs:
       only:
         - master
     docker:
-      - image: node:10
+      - image: node:16
     steps:
       - add_ssh_keys:
           fingerprints:
