+++ package.json
@@ -4,10 +4,13 @@
   "private": true,
   "homepage": "https://bf-v.github.io/bfv",
   "dependencies": {
-    "papaparse": "^5.2.0",
-    "react": "^16.6.3",
-    "react-dom": "^16.6.3",
-    "react-scripts": "2.1.1"
+    "papaparse": "^5.3.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "^5.0.1"
+  },
+  "devDependencies": {
+    "prettier": "^2.7.1"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -16,7 +19,10 @@
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
-    "extends": "react-app"
+    "extends": "react-app",
+    "env": {
+      "es2021": true
+    }
   },
   "browserslist": [
     ">0.2%",
